clean up and fix error bubbling
